requirements.txt
Django~=4.1.3
gunicorn
dj_database_url
psycopg2
requests
django-simpleui
APScheduler
beautifulsoup4~=4.11.1
html2text~=2020.1.16
whitenoise
pyTelegramBotAPI
pathlib~=1.0.1
fake_useragent
undetected_chromedriver
webdriver_manager
more_itertools
packaging
selenium~=4.6.0
binance
python-binance
pybit
python-dateutil
requests_html